requirements.txt
ultralytics
python-telegram-bot==13.13
opencv-python
supervision
python-dotenv